refactor: Content, TodoList 컴포넌트 분리
